refactor: extract ContextEdge to the separate file
