+++ agents/sentiment-analysis/function/handler.py
@@ -20,8 +20,8 @@ def handle(req):
                 "authorUsername": "john_doe",
                 "authorImg": "https://example.com/image.png",
                 "createdAt": "2025-02-19T20:33:29.000Z",
-                "url": "https://twitter.com/john_doe/status/1234567890123456789",
-            },
+                "url": "https://twitter.com/john_doe/status/1234567890123456789"
+            }
         }
     }
     """
