+++ apps/crawler-backend/src/context/entities/context-edge.entity.ts
@@ -1,0 +1,10 @@
+import { Entity, PrimaryColumn } from 'typeorm';
+
+@Entity({ name: 'context_edge' })
+export class ContextEdge {
+  @PrimaryColumn()
+  parent: string;
+
+  @PrimaryColumn()
+  child: string;
+}
